Refactor run command into a separate file
